pyproject.toml
[project]
name = "charla-yingyan"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.68.1",
    "dotenv-python>=0.0.1",
    "gradio>=5.47.2",
    "mcp[cli]>=1.15.0",
    "openai>=1.109.1",
    "pypdf>=6.1.0",
    "python-dotenv>=1.1.1",
]

[tool.uv.workspace]
members = [
    "github-mcp",
]
